cnf ran: rename ztp directory to gitopsztp
Since the ztp directory is already used by assisted, using it in RAN CI requires using labels but this creates a lot of noise as it still runs tests and just skips the ones not selected by the label. By giving RAN ZTP tests the gitopsztp directory name they can still be run using ECO_TEST_FEATURES. Additionally, the test label is also changed to gitopsztp.
