+++ tests/cnf/ran/gitopsztp/internal/tsparams/ztpvars.go
